refactor(utils): remove duplicated function
