+++ packages/docusaurus-plugin-client-redirects/src/collectRedirects.ts
@@ -7,8 +7,8 @@
 
 import _ from 'lodash';
 import logger from '@docusaurus/logger';
-import {addTrailingSlash, removeTrailingSlash} from '@docusaurus/utils';
-import {applyTrailingSlash} from '@docusaurus/utils-common';
+import {removeTrailingSlash} from '@docusaurus/utils';
+import {applyTrailingSlash, addTrailingSlash} from '@docusaurus/utils-common';
 import {
   createFromExtensionsRedirects,
   createToExtensionsRedirects,
