+++ packages/docusaurus-plugin-content-docs/package.json
@@ -41,6 +41,7 @@
     "@docusaurus/module-type-aliases": "3.0.0",
     "@docusaurus/types": "3.0.0",
     "@docusaurus/utils": "3.0.0",
+    "@docusaurus/utils-common": "3.0.0",
     "@docusaurus/utils-validation": "3.0.0",
     "@types/react-router-config": "^5.0.7",
     "combine-promises": "^1.1.0",
