+++ packages/docusaurus-plugin-content-docs/src/sidebars/generator.ts
@@ -8,7 +8,7 @@
 import path from 'path';
 import _ from 'lodash';
 import logger from '@docusaurus/logger';
-import {addTrailingSlash} from '@docusaurus/utils';
+import {addTrailingSlash} from '@docusaurus/utils-common';
 import {createDocsByIdIndex, toCategoryIndexMatcherParam} from '../docs';
 import type {
   SidebarItemDoc,
