+++ packages/docusaurus/package.json
@@ -69,8 +69,8 @@
     "del": "^6.1.1",
     "detect-port": "^1.5.1",
     "escape-html": "^1.0.3",
-    "eval": "^0.1.8",
     "eta": "^2.2.0",
+    "eval": "^0.1.8",
     "file-loader": "^6.2.0",
     "fs-extra": "^11.1.1",
     "html-minifier-terser": "^7.2.0",
